floting point type is changed
float -> double type

+++ common/inc/parSPH_V2/incompressible_sph.h
@@ -10,7 +10,7 @@ public:
 	virtual ~incompressible_sph();
 
 	void setProjectionFrom(tProjectionForm _tproj, size_t order) { tproj = _tproj; }
-	void setPPESolver(size_t _iter, float _tol) { ppeIter = _iter; ppeTol = _tol; }
+	void setPPESolver(size_t _iter, double _tol) { ppeIter = _iter; ppeTol = _tol; }
 
 	virtual bool initialize();
 	virtual void cpuRun();
@@ -21,8 +21,8 @@ private:
 	void predict_the_temporal_velocity();
 	void predict_the_temporal_position();
 	size_t solve_the_pressure_poisson_equation_by_Bi_CGSTAB();
-	void pressure_poisson_equation(float* out, float *p = NULL);
-	void ppe_right_hand_side(float *out);
+	void pressure_poisson_equation(double* out, double *p = NULL);
+	void ppe_right_hand_side(double *out);
 	void correct_by_adding_the_pressure_gradient_term();
 	void particle_shifting();
 	void update_shift_particle();
@@ -34,11 +34,11 @@ private:
 	void predictionStep2();
 	bool solvePressureWithBiCGSTAB();
 	bool solvePressureWithBiCGSTAB2();
-	void PPESolver(float *out, float *pes = NULL);
-	void PPESolver2(float *out, float *pes = NULL);
-	void dummyScalarCopy(float* src);
-	void ghostDummyScalarSet(float* src);
-	float dotProductIgnoreType(float* v1, float* v2, tParticle tp);
+	void PPESolver(double *out, double *pes = NULL);
+	void PPESolver2(double *out, double *pes = NULL);
+	void dummyScalarCopy(double* src);
+	void ghostDummyScalarSet(double* src);
+	double dotProductIgnoreType(double* v1, double* v2, tParticle tp);
 	void correctionStep();
 	void correctionStep2();
 	void first_step();
@@ -48,13 +48,13 @@ private:
 private:
 	tProjectionForm tproj;			// projection from type
 
-	float ppeTol;					// tolerance of poisson pressure equation
+	double ppeTol;					// tolerance of poisson pressure equation
 
 	size_t projOrder;			// projection order
 	size_t ppeIter;			// maximum iteration of poisson pressure equation
 	
-	float* volumes;
-	float* rhs;
+	double* volumes;
+	double* rhs;
 // 	float* lhs;
 // 	float* conjugate0;
 // 	float* conjugate1;
